Add setting symbol timeout
In this [Single Reception Operating] mode, the modem searches for
a preamble during a given period of time. If a preamble hasn’t
been found at the end of the time window, the chip generates the
RxTimeout interrupt and goes back to Standby mode. The length of
the reception window (in symbols) is defined by the RegSymbTimeout
register and should be in the range of 4 (minimum time for the
modem to acquire lock on a preamble) up to 1023 symbols.